Use graph compiler's interface for compile/execute at elyzor backend (#2)

+++ src/graph/backend/elyzor/include/dnnl_graph_compiler.h
@@ -62,7 +62,7 @@ struct dnnl_graph_compiler_context {
 struct dnnl_graph_compiler_tensor {
     size_t id;
     uint8_t ndims;
-    size_t *dims;
+    int64_t *dims;
     void *data;
 };
